Benachmark:  Add library of common functions
